move asyncReplace
